+++ packages/solid-crs-theme/lib/index.ts
@@ -22,6 +22,8 @@ export { default as Logo } from './icons/Logo.svg?raw';
 export { default as Loading } from './icons/Loading.svg?raw';
 export { default as Picture } from './icons/Picture.svg?raw';
 export { default as Reset } from './icons/reset.svg?raw';
+export { default as CheckboxChecked } from './icons/CheckboxChecked.svg?raw';
+export { default as CheckboxUnchecked } from './icons/CheckboxUnchecked.svg?raw';
 
 /**
  * Export theme